# docs/swagger.json
{
	"swagger": "2.0",
	"info": { "contact": {}, "title": "ColaboraTech", "version": "1.0.0" },
	"host": "localhost:8080",
	"basePath": "/",
	"schemes": ["http"],
	"paths": {
...
		"/posts/unlike/{_id}": {
			"parameters": [
				{
					"description": "post Id",
					"in": "path",
					"name": "_id",
					"required": true,
					"type": "string"
				}
			],
			"put": {
				"parameters": [],
				"responses": { "200": { "description": "" } },
				"tags": ["POSTS"],
				"description": "DELETE LIKE",
				"operationId": "deleteLike",
				"summary": "DELETE LIKE"
			}
		},
...
		"/projects/addFavorite/{_id}": {
			"parameters": [
				{ "in": "path", "name": "_id", "required": true, "type": "string" }
			],
			"post": {
				"parameters": [],
				"responses": { "200": { "description": "" } },
				"tags": ["PROJECTS"],
				"description": "ADD FAVORITE",
				"operationId": "addFavorite",
				"summary": "ADD FAVORITE"
			}
		},
		"/projects/removeFavorite/{_id}": {
			"parameters": [
				{ "in": "path", "name": "_id", "required": true, "type": "string" }
			],
			"post": {
				"parameters": [],
				"responses": { "200": { "description": "" } },
				"tags": ["PROJECTS"],
				"description": "REMOVE FAVORITE",
				"operationId": "removeFavorite",
				"summary": "REMOVE FAVORITE"
			}
		},
...
		"/users/name/{name}": {
			"parameters": [
				{
					"description": "user Name",
					"in": "path",
					"name": "name",
					"required": true,
					"type": "string"
				}
			],
			"get": {
				"parameters": [],
				"responses": { "200": { "description": "" } },
				"tags": ["USERS"],
				"description": "Get by name",
				"operationId": "getByName",
				"summary": "Get by name"
			}
		},
...
	},
...
}